distributed logic
